Parse out env for DomainName in cinder-provisioner
When connecting to a deployment using Keystone V3 we
expect/require the Domain Name or ID. This is in the
openrc file for most deployments (ie RDO) however it's
not read in from env variables in the connection code.

This adds the parsing of OS_USER_DOMAIN_NAME to
cinderConfig to fix this and make the provisioner
usable for deployments using Keystone V3.

Fixes Issue: #436