Clean up glide deps
While working on Pull #449 I missed something in the cleaning up the
Glide files.  Not sure if it got lost in the merge, or I dropped it
inadvertently upon commit but regardless I unintentionally added a
ton of gophercloud files to the vendor dir.

This PR is a simple clean up of that.  Ran:
  `glide install -v; glide-vc --use-lock-file`

Should clean up the unnecessary files that were added.