Support noauth connections from the environment
Noauth support was added by a recent commit but it only worked when used
with the config file approach.  Add support for parsing the environment.
If OS_AUTH_URL is set, we will use a Keystone approach.  If
OS_CINDER_ENDPOINT is set we will use noauth.  Add connection tests to
validate that this is working as intended.

Signed-off-by: Adam Litke <alitke@redhat.com>