standalone-cinder: Use cinder attach and detach
Prior to this patch, provisioned cinder volumes were left in 'attaching'
state in the cinder DB.  This is an interim state that is supposed to
resolve to 'attached' or (in the case of failure) 'available'.  This
patch completes the attachment process on the cinder side so that
volumes will appear attached to cinder.

Note that the volume is not really attached to a user until a pod
consumes the PV and the kubelet attaches and mounts the volume.  This
discrepency between k8s and cinder is OK as each system has an
appropriate view of the volume state from its own perspective.

Signed-off-by: Adam Litke <alitke@redhat.com>