standalone-cinder: Pass custom initiatorName
When calling cinder initialize_connection, we pass in a constant
initiatorName.  This value is used by some volume drivers to limit the
exposure of the lun to that specific host.  Since the provisioner
connects the volume before knowing which host will use it we depended on
all hosts in the cluster to be configured with the same matching
initiatorName.  This is not user friendly.  Since kubernetes allows us
to specify a custom initiatorName in the ISCSI PVSource structure we can
avoid the need for the host level configuration and always pass the
constant value.

Signed-off-by: Adam Litke <alitke@redhat.com>