update text on using unique snapshot name
The text was a bit unclear on whether, same snapshot name can be used in a cluster -- on different PV objects or not. Example: is it possible to create:
- Alex creates - Snapshot ( nightly-monday ) on PV - MySQL-VOL
- David creates - Snapshot ( nightly-monday ) on PV - RabbitMQ-VOL

Alex and David run their MySQL and RabbitMQ applications in their own namespaces. 

It appears that - the same name can't be used. Is this right?