[v2] Add block volume support check to external provisioners
This patch add a block volume support check to external provisioners.
This patch introduces BlockProvisioner interface and SupportsBlock()
to check whether the provisioner supports block volume.

They are used by the external provisioner library and if a provisioner
doesn't supports block volume, the library output an error and doesn't try
provisioning the block volume.

With this change, external provisioner that supports block volume
is required to implement SupportsBlock() and return true.
(Also, it must set volumeMode to pv spec that it returns.)