Build efs-provisioner in a multi stage Dockerfile
Using a multi stage Dockerfile makes it slightly easier to build the Docker
image, as we then don't rely on any intermediate state during the build process.

This also fixes a minor bug in the `Dockerfile` as `apk update --no-cache` was
previously, which doesn't have any effect. Updating the package repository index
and then telling `apk` to not store the cache means the command doesn't serve
any purpose.

Instructions on how to build the Docker image has been added to the README file
as well.